split Receive component into coin specific
