+++ src/views/index.js
@@ -18,7 +18,7 @@ import ImportView from 'views/Landing/views/Import/Container';
 import WalletContainer from 'views/Wallet';
 import AccountSummary from 'views/Wallet/views/Account/Summary';
 import AccountSend from 'views/Wallet/views/Account/Send';
-import AccountReceive from 'views/Wallet/views/Account/Receive/Container';
+import AccountReceive from 'views/Wallet/views/Account/Receive';
 import AccountSignVerify from 'views/Wallet/views/Account/SignVerify/Container';
 
 import WalletDashboard from 'views/Wallet/views/Dashboard';
